+++ src/main/java/com/uber/jenkins/phabricator/tasks/SendHarbormasterResultTask.java
@@ -7,18 +7,21 @@ import net.sf.json.JSONNull;
 import net.sf.json.JSONObject;
 
 import java.io.IOException;
+import java.util.Map;
 
 public class SendHarbormasterResultTask extends Task {
 
     private final DifferentialClient diffClient;
     private final String phid;
     private final boolean harbormasterSuccess;
+    private final Map<String, String> coverage;
 
-    public SendHarbormasterResultTask(Logger logger, DifferentialClient diffClient, String phid, boolean harbormasterSuccess) {
+    public SendHarbormasterResultTask(Logger logger, DifferentialClient diffClient, String phid, boolean harbormasterSuccess, Map<String, String> harbormasterCoverage) {
         super(logger);
         this.diffClient = diffClient;
         this.phid = phid;
         this.harbormasterSuccess = harbormasterSuccess;
+        this.coverage = harbormasterCoverage;
     }
 
     /**
@@ -43,21 +46,39 @@ public class SendHarbormasterResultTask extends Task {
     @Override
     protected void execute() {
         try {
-            JSONObject result = diffClient.sendHarbormasterMessage(phid, harbormasterSuccess);
-
-            if (result.containsKey("error_info") && !(result.get("error_info") instanceof JSONNull)) {
-                info(String.format("Error from Harbormaster: %s", result.getString("error_info")));
-                this.result = Result.FAILURE;
-            } else {
-                this.result = Result.SUCCESS;
+            if (!sendMessage(coverage)) {
+                info("Error sending Harbormaster unit results, trying again without unit data (you may have an old Phabricator?).");
+                sendMessage(null);
             }
         } catch (ConduitAPIException e) {
             e.printStackTrace();
-            this.result = Result.FAILURE;
+            failTask();
         } catch (IOException e) {
             e.printStackTrace();
-            this.result = Result.FAILURE;
+            failTask();
         }
+    }
+
+    /**
+     * Try to send a message to harbormaster
+     * @return false if an error was encountered
+     */
+    private boolean sendMessage( Map<String, String> coverage) throws IOException, ConduitAPIException {
+        JSONObject result = diffClient.sendHarbormasterMessage(phid, harbormasterSuccess, coverage);
+
+        if (result.containsKey("error_info") && !(result.get("error_info") instanceof JSONNull)) {
+            info(String.format("Error from Harbormaster: %s", result.getString("error_info")));
+            failTask();
+            return false;
+        } else {
+            this.result = Result.SUCCESS;
+        }
+        return true;
+    }
+
+    private void failTask() {
+        info("Unable to post to Harbormaster");
+        result = result.FAILURE;
     }
 
     /**