+++ src/test/java/com/uber/jenkins/phabricator/conduit/DifferentialClientTest.java
@@ -111,7 +111,7 @@ public class DifferentialClientTest {
     public void testSendHarbormasterSuccess() throws IOException, ConduitAPIException {
         JSONObject empty = new JSONObject();
         mockConduitResponse(differentialClient, empty);
-        differentialClient.sendHarbormasterMessage(TestUtils.TEST_PHID, true);
+        differentialClient.sendHarbormasterMessage(TestUtils.TEST_PHID, true, null);
     }
 
     private void mockConduitResponse(DifferentialClient client, JSONObject response) throws IOException, ConduitAPIException {